inprove a bit the aspect of string objects

+++ ummisco.gama.opengl/src/ummisco/gama/modernOpenGL/font/fontMeshCreator/FontTextureCache.java
@@ -12,12 +12,12 @@ public class FontTextureCache {
 	
 	private HashMap<String,Texture> textureMap = new HashMap<String,Texture>();
 	
-	public TextMeshData getTextMeshData(String textureName, String content, int yRatioBetweenPixelsAndModelUnits, int textSize) {
+	public TextMeshData getTextMeshData(String textureName, String content, double yRatioBetweenPixelsAndModelUnits, int textSize) {
 		String absolutePathToFontTextureFolder = this.getClass().getProtectionDomain().getCodeSource().getLocation().getPath() 
 				+ "res" + File.separator + "font" + File.separator;
 		String fontFile = absolutePathToFontTextureFolder + textureName + ".fnt";
 		FontType font = new FontType(fontFile);
-		float scale = (10f / yRatioBetweenPixelsAndModelUnits) * textSize;
+		float scale = (10f / (float)yRatioBetweenPixelsAndModelUnits) * textSize;
 		GUIText text = new GUIText(content, scale, font, -1, true);
 		return font.loadText(text);
 	}
